added token valiation, light of format

+++ src/main/java/uk/casey/cryptodash/WatchListController.java
@@ -5,11 +5,15 @@ import java.sql.Timestamp;
 import java.time.Instant;
 import java.util.List;
 import java.util.UUID;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.lang.NonNull;
 import org.springframework.web.bind.annotation.*;
 import uk.casey.cryptodash.models.WatchListItemModel;
 import uk.casey.cryptodash.services.WatchListService;
+import uk.casey.cryptodash.utils.JwtUtil;
 
 @RestController
 public class WatchListController {
@@ -20,9 +24,17 @@ public class WatchListController {
     this.watchListService = watchListService;
   }
 
+  private static final Logger logger = LoggerFactory.getLogger(WatchListController.class);
+
   @GetMapping("/api/user/watchlist")
   public ResponseEntity<List<WatchListItemModel>> getWatchList(
-      @RequestParam String fiat, @RequestHeader("userId") UUID userId) {
+      @RequestParam String fiat,
+      @RequestHeader("userId") UUID userId,
+      @RequestHeader("authorisation") String token) {
+
+    if (!JwtUtil.isValidJwtFormat(token)) {
+      return ResponseEntity.status(401).build();
+    }
 
     List<WatchListItemModel> watchlist = watchListService.getWatchList(userId);
 
@@ -33,8 +45,13 @@ public class WatchListController {
   public ResponseEntity<Boolean> addToWatchlist(
       @RequestParam String name,
       @RequestHeader("userId") UUID userId,
+      @RequestHeader("authorisation") String token,
       @RequestBody WatchListItemModel req)
       throws SQLException {
+
+    if (!JwtUtil.isValidJwtFormat(token)) {
+      return ResponseEntity.status(401).build();
+    }
 
     watchListService.addToWatchlist(
         userId,
@@ -50,7 +67,13 @@ public class WatchListController {
 
   @DeleteMapping("/api/user/watchlist")
   public ResponseEntity<Boolean> deleteWatchlist(
-      @RequestParam int assetId, @RequestHeader("userId") UUID userId) {
+      @RequestParam int assetId,
+      @RequestHeader("userId") UUID userId,
+      @RequestHeader("authorisation") String token) {
+
+    if (!JwtUtil.isValidJwtFormat(token)) {
+      return ResponseEntity.status(401).build();
+    }
 
     watchListService.deleteFromWatchlist(userId, assetId);
 
@@ -61,7 +84,12 @@ public class WatchListController {
   public ResponseEntity<Boolean> updateWatchlist(
       @RequestParam int assetId,
       @RequestHeader("userId") UUID userId,
+      @RequestHeader("authorisation") @NonNull String token,
       @RequestBody WatchListItemModel req) {
+
+    if (!JwtUtil.isValidJwtFormat(token)) {
+      return ResponseEntity.status(401).build();
+    }
 
     watchListService.updateItemWatchList(userId, assetId, req.getValue());
